pyproject.toml
[project]
name = "podium-hackathon"
version = "0.1.0"
description = "Autonomous business agents powered by LangGraph"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4",
    "langchain-core>=0.3",
    "langchain-anthropic>=0.3",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
]
